+++ play_scraper/__init__.py
@@ -6,6 +6,6 @@ Google Play Store Scraper
 A web scraper for the Google Play Android app store.
 """
 
-__version__ = '0.1.1'
+__version__ = '0.1.2'
 
 from .api import details, collection, developer, suggestions, search, similar