+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Changelog
 
-### 0.1.2 - 2016-04-17
+### 0.1.3: 2016-04-20
+
+* Python3 urllib.parse.quote_plus compatibility in scraper/utils
+* Python3 import fixes
+* No reviews (ratings) set to 0 instead of None
+* Developer ID included when available
+
+### 0.1.2: 2016-04-17
 
 * Bugfix: AGE_RANGE params for fetching collections in the FAMILY category
 * Bugfix: description_html doesn't include parent div  anymore
 * README examples updates
 
-### 0.1.1 - 2016-04-17
+### 0.1.1: 2016-04-17
 
 * Added urljoin import compatibility for Python 3
 * Simplified parsing description to just use unicode